Removed warning messages.

+++ middleware/wifi/dhcpd/dhcp-server.c
@@ -259,7 +259,7 @@ int dhcp_send_response(int sock, struct sockaddr *addr, char *msg, int len)
 {
     int nb;
     unsigned int sent = 0;
-    while (sent < len)
+    while (sent < (unsigned int)len)
     {
         nb = sendto(sock, msg + sent, len - sent, 0, addr, sizeof(struct sockaddr_in));
         if (nb < 0)
@@ -286,7 +286,7 @@ static int process_dhcp_message(char *msg, int len)
     bool got_client_ip    = 0;
     uint32_t new_ip;
 
-    if (!msg || len < sizeof(struct bootp_header) + sizeof(struct bootp_option) + 1)
+    if (!msg || (unsigned int)len < sizeof(struct bootp_header) + sizeof(struct bootp_option) + 1)
         return -WM_E_DHCPD_INVALID_INPUT;
 
     hdr = (struct bootp_header *)msg;
